refactor: adjust database package module
