+++ database/migration/migration.go
@@ -1,4 +1,4 @@
-package database
+package migration
 
 import (
 	"senkou-catalyst-be/app/models"
